requirements.txt
flickrapi==2.1.2
pillow==3.2.0
requests==2.10.0
tweebot==0.2.1
